pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "A professional voice agent built with LiveKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-api>=1.1.0",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-openai",
    "livekit-plugins-deepgram",
    "livekit-plugins-assemblyai",
    "livekit-plugins-cartesia",
    "livekit-plugins-ai-coustics",
    "livekit-plugins-noise-cancellation~=0.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv",
    "torch",
    "mcp[cli]",
    "asyncpg",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mantra*"]
exclude = ["static*", "recordings*", "massist*"]

[project.scripts]
mantra-agent = "mantra.agent:run_agent"
mantra-ui = "mantra.ui_server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
